Refactored the GameWindowViewModels
Now the GameWindowViewModel is responsable for all game logic. The BoardViewModel is no more.
