+++ TicTacToeGameCompose/app/build.gradle
@@ -50,6 +50,10 @@ android {
 
 dependencies {
 
+    def nav_version = "2.7.3"
+
+    implementation "androidx.navigation:navigation-compose:$nav_version"
+
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation platform('org.jetbrains.kotlin:kotlin-bom:1.8.0')
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.2'
